requirements.txt
alabaster==1.0.0
asgiref==3.8.1
attrs==24.2.0
babel==2.16.0
cattrs==24.1.2
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
docutils==0.21.2
drf-yasg==1.21.8
esbonio==0.16.5
exceptiongroup==1.2.2
Faker==33.1.0
idna==3.10
imagesize==1.4.1
inflection==0.5.1
Jinja2==3.1.4
lsprotocol==2023.0.1
MarkupSafe==3.0.2
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
pygls==1.3.1
Pygments==2.18.0
pyspellchecker==0.8.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
snowballstemmer==2.2.0
Sphinx==8.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.2
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
